backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0

